Pan_Auto_fill/backend/requirements.txt
flask
flask-cors
opencv-python
numpy
pytesseract
ultralytics
mysql-connector-python
